Update crud operations
Added a check for the existence of a user in the get_user_by_id function before attempting to retrieve their data.

The get_filtered_users function now accepts an additional parameter, letter, to filter users by a specific letter in the username.

Checks for the existence of a user have been added in the update_user_by_id and delete_user_by_id functions before performing update and delete operations, respectively.